Implemented global wind speed and direction for the 2D and cumulus clouds.
Adds a "Weather" section to the Sky3D node with properties to set both the wind speed (in m/s) and direction (in degrees, measured from the north). Changing these will set the appropriate direction and speed for both the 2D and cumulus clouds in the sky dome.
Also makes a few changes to the sky material shader to allow smooth transitions in the cloud animations when changing either the wind speed or direction. Before, clouds would "skip" whenever you made such changes.

Ensured optimal precision for 2D cloud position by wrapping the position around after it reaches 1.0. This same concept does not work for the cumulus clouds, so if you have strong winds in the same direction for a long time, those might still start "stuttering".